Replace market daemon with single avanza_market_multi module
Old market daemon runs multiple processes for different stocks at the same time.
Avanza kicks you out because of too frequent login attempt and lock account temporarily.
This is because each process initiates an Avanza instance, and all instances log in at the same time.

We only need one Avanza instance to listen to all different SSE pushes.

Note that the chart app is not yet updated with this new module in this commit.